inference/requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
-f https://data.pyg.org/whl/torch-2.0.0+cpu.html
torch
torchvision
diffusers
gradio
Pillow
numpy
transformers
torchvision
ftfy
altair
opencv-python
accelerate